requirements.txt
Django==4.1.2
PyJWT==2.5.0
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
kafka-python==2.0.2
pytest==7.1.3
python-dotenv==0.21.0
requests==2.28.1
